+++ clock/clock/main.cpp
@@ -1,0 +1,9 @@
+#include "stdafx.h"
+#include "sheet.h"
+#include "clock.cpp"
+
+int main(void)
+{
+	enterTimeLoop();
+	return 0;
+}
